+++ src/daft-plan/src/physical_ops/mod.rs
@@ -1,5 +1,6 @@
 mod agg;
 mod csv;
+mod fanout;
 mod filter;
 #[cfg(feature = "python")]
 mod in_memory;
@@ -12,6 +13,7 @@ mod split;
 
 pub use agg::Aggregate;
 pub use csv::TabularScanCsv;
+pub use fanout::{FanoutByHash, FanoutByRange, FanoutRandom};
 pub use filter::Filter;
 #[cfg(feature = "python")]
 pub use in_memory::InMemoryScan;
@@ -20,4 +22,4 @@ pub use limit::Limit;
 pub use parquet::TabularScanParquet;
 pub use reduce::ReduceMerge;
 pub use sort::Sort;
-pub use split::{Split, SplitByHash, SplitByRange, SplitRandom};
+pub use split::Split;
