+++ src/daft-plan/src/physical_ops/split.rs
@@ -1,7 +1,5 @@
 use std::sync::Arc;
 
-use daft_dsl::Expr;
-
 use crate::physical_plan::PhysicalPlan;
 
 #[derive(Clone, Debug)]
@@ -21,63 +19,6 @@ impl Split {
         Self {
             input_num_partitions,
             output_num_partitions,
-            input,
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct SplitRandom {
-    pub num_partitions: usize,
-    // Upstream node.
-    pub input: Arc<PhysicalPlan>,
-}
-
-impl SplitRandom {
-    pub(crate) fn new(num_partitions: usize, input: Arc<PhysicalPlan>) -> Self {
-        Self {
-            num_partitions,
-            input,
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct SplitByHash {
-    pub num_partitions: usize,
-    pub partition_by: Vec<Expr>,
-    // Upstream node.
-    pub input: Arc<PhysicalPlan>,
-}
-
-impl SplitByHash {
-    pub(crate) fn new(
-        num_partitions: usize,
-        partition_by: Vec<Expr>,
-        input: Arc<PhysicalPlan>,
-    ) -> Self {
-        Self {
-            num_partitions,
-            partition_by,
-            input,
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct SplitByRange {
-    pub num_partitions: usize,
-    pub sort_by: Vec<Expr>,
-    // Upstream node.
-    pub input: Arc<PhysicalPlan>,
-}
-
-impl SplitByRange {
-    #[allow(dead_code)]
-    pub(crate) fn new(num_partitions: usize, sort_by: Vec<Expr>, input: Arc<PhysicalPlan>) -> Self {
-        Self {
-            num_partitions,
-            sort_by,
             input,
         }
     }
